+++ src/Main/HtmlMapper/DocsToHtmlMapper.ts
@@ -1,10 +1,10 @@
 import { List } from "../../Utility/List"
-import { HtmlCreatorMapper } from "../HtmlCreator"
-import { Document } from "./Document"
-import { Html_Image } from "./Image/Html_Image"
-import { ImageMapper } from "./Image/Mapper"
-import { ParagraphElement } from "./ParagraphElements"
-import { TextRunMapper } from "./TextRun/Mapper"
+import { HtmlCreatorMapper } from "./HtmlCreator"
+import { Document } from "../InterfaceLayer/Document"
+import { Html_Image } from "./Html_Image"
+import { ImageMapper } from "./ImageMapper"
+import { ParagraphElement } from "../InterfaceLayer/ParagraphElements"
+import { TextRunMapper } from "./TextRunMapper"
 
 function concatStrings(str1:string, str2:string):string{
     return str1.concat(str2)
@@ -15,7 +15,7 @@ export class HtmlMapper3{
         return doc.compactMap(paragraphElementsToString).reduce(listToParagraph, " ").trim() 
 
         function listToParagraph(htmlString:string, list:List<string>):string{
-            return htmlString.concat( HtmlCreatorMapper.wrapInParagraphTag(list.reduce(concatStrings)) )
+            return htmlString.concat( HtmlCreatorMapper.createTag("p", list.reduce(concatStrings)) )
         } 
         function paragraphElementsToString(element: ParagraphElement): string{
             return element.type == "image" ? ImageMapper.toHtml(Html_Image.from(element)) : TextRunMapper.createHtml(element)
