+++ src/Main/InterfaceLayer/ParagraphElements.ts
@@ -1,11 +1,11 @@
 import { DocsInlineObjectsModel, DocsParagraphElementModel } from "../../Models/DocsDocumentModel"
-import { Maybe, orElseGet } from "../../Utility/Maybe"
-import { Image, IImage } from "./Image/ImageObj"
-import { TextRun, ITextRun } from "./TextRun/TextRun"
+import { Maybe, MaybeUtility } from "../../Utility/Maybe"
+import { Image, IImage } from "./ImageObj"
+import { TextRun, ITextRun } from "./TextRun"
 
 export class ParagraphElementMaker{
     static from( element:DocsParagraphElementModel, inlineObj?:DocsInlineObjectsModel ): Maybe<ParagraphElement>{
-        return orElseGet(
+        return MaybeUtility.orElseGet(
             () => Image.from(element?.inlineObjectElement, inlineObj) as Maybe<ParagraphElement>,
             () => TextRun.from(element?.textRun) as Maybe<ParagraphElement>
         )()
